farmer_chat-master/backend/requirements.txt
flask
python-dotenv
openai
deep-translator
textblob
flask-cors
gTTS
sqlalchemy
flask-sqlalchemy
deepgram-sdk
requests
openai-whisper
transformers
torch
sentencepiece